Fix argument parse bug.
When large number was passed, program abort with exception

+++ loop/loop.cpp
@@ -3,7 +3,7 @@
 
 int main(int argc, char *argv[]) {
     auto arg = argv[1];
-    int64_t upper = std::stoi(arg);
+    int64_t upper = std::stoll(arg);
 
     int64_t sum = 0;
     for (int64_t i = 0; i <= upper; ++i) {
